Fixed tests failing with session transactions and ran linting. MODIFIED > src/views/feedback.py > tests/__init__.py > tests/test_api.py > tests/test_feedback.py REMOVED > tests/conftest.py
